requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.92.2
